refactor delete user, candidate, post, comment and their tests

+++ src/services/user.service.ts
@@ -1,5 +1,6 @@
 import type { UpdateUserForm } from '../types/user.type'
 import { User } from '../models/user.model'
+import { deleteCommentById } from './comment.service'
 
 export const findUserById = async (id: string) => {
   return await User.findById(id)
@@ -22,5 +23,14 @@ export const getAllUsers = async () => {
 }
 
 export const deleteUserById = async (id: string) => {
+  const user = await findUserById(id)
+  if (!user) return user
+
+  const comments = user.comments
+  const deleteCommentPromises = comments.map(async (comment) => {
+    await deleteCommentById(comment._id.toString())
+  })
+  await Promise.all(deleteCommentPromises)
+
   return await User.findByIdAndDelete(id)
 }
